backend2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
better-exceptions==0.3.2
openai==1.3.8
qdrant-client==1.9.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
python-dotenv==1.0.0
aiosqlite==0.19.0
httpx==0.25.2
limits==3.9.0
aioredis==2.0.1
google-generativeai==0.4.1
tensorflow